requirements.txt
astroid==2.4.2
click==7.1.2
colorama==0.4.3
Flask==1.1.2
gunicorn==20.0.4
isort==5.5.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.6.0
pytube3==9.6.4
six==1.15.0
toml==0.10.1
typing-extensions==3.7.4.3
Werkzeug==1.0.1
wrapt==1.12.1
